llm_compression_poc/python/requirements.txt
transformers
torch
coremltools
accelerate
exporters
python-dotenv